+++ src/Fields/Wysiwyg.php
@@ -10,20 +10,9 @@ class Wysiwyg extends BaseField
     protected $name = 'content';
     protected $label = 'Inhalt';
     protected $type = 'wysiwyg';
-
-    /**
-     * @param array  $parameter
-     *
-     * @return array
-     */
-    public function build(array $parameter = [])
-    {
-        $add_to_defaults = [
-            'tabs' => 'all',
-            'toolbar' => 'full',
-            'media_upload'  => 0,
-        ];
-
-        return parent::build(array_merge((array) $add_to_defaults, (array) $parameter));
-    }
+    protected $args = [
+        'tabs' => 'all',
+        'toolbar' => 'full',
+        'media_upload'  => 0,
+    ];
 }